+++ neurolabi/gui/zcubearray.h
@@ -34,7 +34,7 @@ public:
     std::vector<glm::vec4> colors;
     std::vector<glm::vec3> normals;
     std::vector<glm::vec2> texCoords;
-    std::vector<int> normalIndices;
+    std::vector<float> normalIndices;
 
     std::vector<glm::vec3> ePositions;
     std::vector<glm::vec4> eColors;
@@ -135,7 +135,7 @@ public:
 
     std::vector<glm::vec3> norm;
     std::vector<glm::vec2> texc;
-    std::vector<int> normIndex;
+    std::vector<float> normIndex;
 };
 
 //
